+++ example/lib/components/custom_scroll_behavior.dart
@@ -1,0 +1,15 @@
+import 'dart:ui';
+import 'package:flutter/material.dart';
+
+class CustomScrollBehavior extends MaterialScrollBehavior {
+  const CustomScrollBehavior();
+
+  @override
+  Set<PointerDeviceKind> get dragDevices {
+    return const {
+      PointerDeviceKind.touch,
+      PointerDeviceKind.mouse,
+      PointerDeviceKind.trackpad,
+    };
+  }
+}