+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # 0.0.1-alpha.1
 
-- The basic anaglyph implementation
+- Basic anaglyph implementation
 
 # 0.0.1-alpha.2
 
 - pubspec description lengthened to increase the pub score
+
+# 0.0.1-beta
+
+- `debugPaintSize()` has been implemented for `RenderAnaglyphView`
+- `performLayout()` has been implemented for `RenderAnaglyphView`
+- A fancy animated page has been added to the example
+- Dynamic size anaglyph tutorial has been added to the docs